+++ src/Common/StringTools.h
@@ -117,6 +117,6 @@ bool parseIpAddressAndPort(uint32_t& ip, uint32_t& port, const std::string& addr
 
 std::string timeIntervalToString(uint64_t intervalInSeconds);
 
-std::string repeatChar(int count, char c);
+std::string repeatChar(size_t count, char c);
 
 }
